requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
torch==2.2.2
simple-lama-inpainting==0.1.2
Pillow==10.3.0
